+++ 1.21/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.0.28-1.21] - 2024-09-08
+### Changed
+- Allow for preliminary compatibility with Minecraft 1.21.1 when enabled via Fabric's [Dependency Overrides](https://fabricmc.net/wiki/tutorial:dependency_overrides)
+- Update mixin implementation for `AnimalTameCallback` and `ProjectileImpactCallback` on Fabric
+
 ## [v21.0.27-1.21] - 2024-09-06
 ### Added
 - Experimental settings screens are now automatically skipped to avoid being shown for bundled data packs